requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.5
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
flake8==6.0.0
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
inflection==0.5.1
jsonschema==4.19.0
mccabe==0.7.0
pep8-naming==0.15.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pycodestyle==2.10.0
pyflakes==3.0.0
PyJWT==2.8.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0.2
sqlparse==0.4.4
uritemplate==4.1.1
celery>=5.2
redis>=4.0
